Make fetch and render independent via a shared common module
fetch.py imported Cache, update_from_web, extract_derived and load_ctx
from render.py, coupling the two stages. Move the fetch-only logic into
fetch.py and extract the genuinely shared helpers (load_ctx, reduce_sc,
strip_html, SMALL_CAPS) into common.py, so neither script imports the
other and nothing is duplicated.
